rename utilFuncs -> utils
